Added AStarTest and Test concept. Bug fixes.
Added JUnit 4.12. You have to download it and add it, so the Unit tests work :).

+++ src/com/hambbe/graph/DiGenericGraph.java
@@ -4,7 +4,7 @@ import java.util.function.Function;
 
 public class DiGenericGraph<V, K> extends GenericGraph<V, K> {
 
-    public DiGenericGraph(Function<K, Integer> edgeToWeight) { super(edgeToWeight); }
+    public DiGenericGraph(Function<K, Double> edgeToWeight) { super(edgeToWeight); }
 
     public void connect(Item v1, Item v2, K edgeValue) {
         checkMembership(v1, v2);
